Add append to file output streams.  Construct J2ME basics for resumable download

+++ core/src/com/ustadmobile/core/impl/UMTransferJob.java
@@ -54,7 +54,7 @@ public interface UMTransferJob {
      * 
      * @return Number of bytes processed so far
      */
-    public int getBytesDownloadedCount();
+    public long getBytesDownloadedCount();
     
     /**
      * Get the total size of the job in bytes or -1 if unknown
